# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/collections": {
      "post": {
        "description": "Creates a collection in the caller's pinned namespace (the `x-beeable-namespace` header, set by the SDK). The namespace must already exist — call `POST /api/namespaces/ensure` on the platform first.",
        "operationId": "createCollection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "schemaVersion": {
                    "type": "integer",
                    "default": 1
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "namespaceId",
                    "name",
                    "schemaVersion",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "namespaceId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "schemaVersion": {
                      "type": "integer"
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Create a named collection",
        "tags": [
          "collections"
        ]
      },
      "get": {
        "description": "When `x-beeable-namespace` is pinned, lists collections in that one namespace. Otherwise lists across every namespace the caller can access.",
        "operationId": "listCollections",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of collections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "namespaceId",
                      "name",
                      "schemaVersion",
                      "metadata",
                      "createdAt",
                      "updatedAt"
                    ],
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "namespaceId": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "name": {
                        "type": "string"
                      },
                      "schemaVersion": {
                        "type": "integer"
                      },
                      "metadata": {
                        "type": "object",
                        "additionalProperties": true
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "List collections in the caller's namespaces",
        "tags": [
          "collections"
        ]
      }
    },
    "/api/collections/ensure": {
      "post": {
        "description": "Returns the existing collection when one exists with the same name in the pinned namespace; otherwise creates it. `created` indicates which branch ran.",
        "operationId": "ensureCollection",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "schemaVersion": {
                    "type": "integer",
                    "default": 1
                  },
                  "metadata": {
                    "type": "object",
                    "additionalProperties": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "collection",
                    "created"
                  ],
                  "properties": {
                    "collection": {
                      "type": "object",
                      "required": [
                        "id",
                        "namespaceId",
                        "name",
                        "schemaVersion",
                        "metadata",
                        "createdAt",
                        "updatedAt"
                      ],
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "namespaceId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string"
                        },
                        "schemaVersion": {
                          "type": "integer"
                        },
                        "metadata": {
                          "type": "object",
                          "additionalProperties": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "created": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Idempotently create-or-fetch a collection in the pinned namespace",
        "tags": [
          "collections"
        ]
      }
    },
    "/api/collections/{name}": {
      "get": {
        "operationId": "getCollection",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Collection",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "namespaceId",
                    "name",
                    "schemaVersion",
                    "metadata",
                    "createdAt",
                    "updatedAt"
                  ],
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "namespaceId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "name": {
                      "type": "string"
                    },
                    "schemaVersion": {
                      "type": "integer"
                    },
                    "metadata": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get a single collection by name",
        "tags": [
          "collections"
        ]
      },
      "delete": {
        "operationId": "deleteCollection",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "summary": "Delete a collection and every document inside it",
        "tags": [
          "collections"
        ]
      }
    },
...
  },
  "info": {
    "title": "Beeable Data Daemon",
    "description": "Generic named JSON-document collections backed by Postgres JSONB",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}